add derive & schemars features
