+++ rust/examples/mcap/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 workspace = true
 
 [dependencies]
-foxglove = { path = "../../foxglove", default-features = false }
+foxglove = { path = "../../foxglove", default-features = false, features = ["derive", "schemars"] }
 clap = { version = "4.5", features = ["derive"] }
 ctrlc = "3.4.5"
 env_logger = "0.11.5"
