+++ rust/foxglove/src/lib.rs
@@ -287,6 +287,8 @@ pub mod log_macro;
 mod log_sink_set;
 mod mcap_writer;
 mod metadata;
+#[doc(hidden)]
+#[cfg(feature = "derive")]
 pub mod protobuf;
 mod schema;
 pub mod schemas;
@@ -325,10 +327,11 @@ pub use runtime::shutdown_runtime;
 #[cfg(feature = "live_visualization")]
 pub use websocket_server::{WebSocketServer, WebSocketServerHandle};
 
-extern crate foxglove_derive;
 #[doc(hidden)]
+#[cfg(feature = "derive")]
 pub use foxglove_derive::Encode;
 #[doc(hidden)]
+#[cfg(feature = "derive")]
 pub use prost_types;
 
 /// An error type for errors generated by this crate.
